requirements.txt
uvicorn==0.23.0
fastapi==0.100.0
psycopg2-binary